Lottie animation added

+++ app/src/main/java/gustavo/projects/pinweathermap/network/response/GetWeatherByCoordResponse.kt
@@ -5,7 +5,7 @@ data class GetWeatherByCoordResponse(
         val clouds: CloudsDto,
         val cod: Int,
         val coord: CoordDto,
-        val dt: Int,
+        val dt: Long,
         val id: Int,
         val main: MainDto,
         val name: String,
